refactor(error-boundary): improve consistency of package
