+++ packages/react/error-boundary/src/ErrorBoundary.spec.tsx
@@ -1,6 +1,6 @@
-import { act, render, screen } from '@testing-library/react';
+import { act, render, renderHook, screen } from '@testing-library/react';
 import { ComponentRef, createRef } from 'react';
-import ErrorBoundary from './ErrorBoundary';
+import { ErrorBoundary, useErrorBoundary } from './ErrorBoundary';
 
 const TEXT_ERROR = `This is an error`;
 const TEXT_NO_ERROR = `This is no error`;
@@ -44,5 +44,34 @@ describe('ErrorBoundary', () => {
     });
 
     expect(screen.getByText(TEXT_NO_ERROR)).toBeInTheDocument();
+  });
+});
+
+describe('useErrorBoundary', () => {
+  it('반환되는 setError에 Error를 넘겨주면 throw 된다.', async () => {
+    // Given
+    const error = new Error('test');
+
+    // When
+    const { result } = renderHook(useErrorBoundary, {
+      wrapper: ({ children }) => {
+        return (
+          <ErrorBoundary
+            renderFallback={({ error }) => {
+              return <>{error.message}</>;
+            }}
+          >
+            {children}
+          </ErrorBoundary>
+        );
+      },
+    });
+
+    act(() => {
+      result.current(error);
+    });
+
+    // Then
+    expect(await screen.findByText('test')).toBeInTheDocument();
   });
 });
