+++ packages/react/error-boundary/src/ErrorBoundary.tsx
@@ -3,6 +3,7 @@ import { isDifferentArray } from '@toss/utils';
 import {
   Component,
   ComponentPropsWithoutRef,
+  ComponentType,
   ErrorInfo,
   forwardRef,
   PropsWithChildren,
@@ -11,8 +12,11 @@ import {
   useContext,
   useImperativeHandle,
   useRef,
+  useState,
 } from 'react';
 import { ErrorBoundaryGroupContext } from './ErrorBoundaryGroup';
+import { ComponentPropsWithoutChildren } from './types';
+import { isDevelopment } from './utils';
 
 type RenderFallbackProps<ErrorType extends Error = Error> = {
   error: ErrorType;
@@ -22,28 +26,31 @@ type RenderFallbackProps<ErrorType extends Error = Error> = {
 type RenderFallbackType = <ErrorType extends Error>(props: RenderFallbackProps<ErrorType>) => ReactNode;
 type IgnoreErrorType = <ErrorType extends Error = Error>(error: ErrorType) => boolean;
 
-type Props<ErrorType extends Error = Error> = {
-  /**
-   * @description 발생할 수 있는 error에 대한 기준값으로 이 값이 변경되면 error를 초기화합니다.
-   */
-  resetKeys?: unknown[];
-  onReset?(): void;
-  renderFallback: RenderFallbackType;
-  onError?(error: ErrorType, info: ErrorInfo): void;
-  /**
-   * @description 이 ErrorBoundary Context에서 처리하지 않고 throw해줄 error의 조건을 명시할 callback
-   */
-  ignoreError?: IgnoreErrorType;
-};
+type ErrorBoundaryProps<ErrorType extends Error = Error> = PropsWithRef<
+  PropsWithChildren<{
+    /**
+     * @description 발생할 수 있는 error에 대한 기준값으로 이 값이 변경되면 error를 초기화합니다.
+     */
+    resetKeys?: unknown[];
+    onReset?(): void;
+    renderFallback: RenderFallbackType;
+    onError?(error: ErrorType, info: ErrorInfo): void;
+    /**
+     * @description 이 ErrorBoundary Context에서 처리하지 않고 throw해줄 error의 조건을 명시할 callback
+     */
+    ignoreError?: IgnoreErrorType;
+  }>
+>;
 
-interface State<ErrorType extends Error = Error> {
+interface ErrorBoundaryState<ErrorType extends Error = Error> {
   error: ErrorType | null;
 }
 
-const initialState: State = {
+const initialState: ErrorBoundaryState = {
   error: null,
 };
-class BaseErrorBoundary extends Component<PropsWithRef<PropsWithChildren<Props>>, State> {
+
+class BaseErrorBoundary extends Component<ErrorBoundaryProps, ErrorBoundaryState> {
   state = initialState;
   updatedWithError = false;
 
@@ -71,7 +78,7 @@ class BaseErrorBoundary extends Component<PropsWithRef<PropsWithChildren<Props>>
     this.resetState();
   };
 
-  componentDidUpdate(prevProps: Props) {
+  componentDidUpdate(prevProps: ErrorBoundaryProps) {
     const { error } = this.state;
 
     if (error == null) {
@@ -105,7 +112,7 @@ class BaseErrorBoundary extends Component<PropsWithRef<PropsWithChildren<Props>>
   }
 }
 
-const ErrorBoundary = forwardRef<{ reset(): void }, ComponentPropsWithoutRef<typeof BaseErrorBoundary>>(
+export const ErrorBoundary = forwardRef<{ reset(): void }, ComponentPropsWithoutRef<typeof BaseErrorBoundary>>(
   (props, resetRef) => {
     const group = useContext(ErrorBoundaryGroupContext);
     const resetKeys = group.resetKey ? [group.resetKey, ...(props.resetKeys || [])] : props.resetKeys;
@@ -119,4 +126,30 @@ const ErrorBoundary = forwardRef<{ reset(): void }, ComponentPropsWithoutRef<typ
   }
 );
 
-export default ErrorBoundary;
+export const useErrorBoundary = <ErrorType extends Error>() => {
+  const [error, setError] = useState<ErrorType | null>(null);
+
+  if (error != null) {
+    throw error;
+  }
+
+  return setError;
+};
+
+export const withErrorBoundary = <Props extends Record<string, unknown> = Record<string, never>>(
+  Component: ComponentType<Props>,
+  errorBoundaryProps: ComponentPropsWithoutChildren<typeof ErrorBoundary>
+) => {
+  const Wrapped = (props: Props) => (
+    <ErrorBoundary {...errorBoundaryProps}>
+      <Component {...props} />
+    </ErrorBoundary>
+  );
+
+  if (isDevelopment) {
+    const name = Component.displayName || Component.name || 'Component';
+    Wrapped.displayName = `withErrorBoundary(${name})`;
+  }
+
+  return Wrapped;
+};
