+++ packages/react/error-boundary/src/index.ts
@@ -1,5 +1,3 @@
 /** @tossdocs-ignore */
-export { default as ErrorBoundary } from './ErrorBoundary';
+export { ErrorBoundary, useErrorBoundary, withErrorBoundary } from './ErrorBoundary';
 export { ErrorBoundaryGroup, useErrorBoundaryGroup, withErrorBoundaryGroup } from './ErrorBoundaryGroup';
-export { default as useErrorBoundary } from './useErrorBoundary';
-export { default as withErrorBoundary } from './withErrorBoundary';
