+++ packages/react/error-boundary/src/types/index.ts
@@ -1,0 +1,5 @@
+/** @tossdocs-ignore */
+import { ComponentProps, JSXElementConstructor } from 'react';
+
+export type ComponentPropsWithoutChildren<Component extends keyof JSX.IntrinsicElements | JSXElementConstructor<any>> =
+  Omit<ComponentProps<Component>, 'children'>;
