--- packages/react/error-boundary/src/useErrorBoundary.ts
@@ -1,12 +1,0 @@
-/** @tossdocs-ignore */
-import { useState } from 'react';
-
-export default function useErrorBoundary<ErrorType extends Error>() {
-  const [error, setError] = useState<ErrorType | null>(null);
-
-  if (error != null) {
-    throw error;
-  }
-
-  return setError;
-}
