+++ packages/react/error-boundary/src/utils/index.ts
@@ -1,0 +1,2 @@
+/** @tossdocs-ignore */
+export const isDevelopment = process.env.NODE_ENV !== 'production';
